LayoutTests/inspector/debugger/scripts-panel-expected.txt

Tests that scripts panel UI elements work as intended.

Debugger was enabled.

Running: testInitialLoad

Running: testHistory
Debugger was disabled.
